requirements.txt
# Core
pytz==2021.3
Django==4.0.3
django-configurations==2.3.2
gunicorn==20.1.0
newrelic==7.6.0.173

# For the persistence stores
psycopg2-binary==2.9.3
dj-database-url==0.5.0

# Model Tools
django-model-utils==4.2.0
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.13.1
Markdown==3.3.6
django-filter==21.1

# Developer Tools
ipdb==0.13.9
ipython==8.1.1
mkdocs==1.2.3
flake8==4.0.1

# Testing
mock==4.0.3
factory-boy==3.2.1
django-nose==1.4.7
nose-progressive==1.5.2
coverage==6.3.2

# Static and Media Storage
django-storages==1.12.3
boto3==1.21.13

# Open API
django-rest-swagger==2.2.0
drf-yasg==1.20.0
